work on edit feature, make modal into component
